+++ packages/client/src/pages/EmissionsMetricsPage/CarbonComparisonCard/carbonComparisonStyles.ts
@@ -3,8 +3,9 @@
  */
 
 import { makeStyles } from '@material-ui/core/styles'
+import { CCFTheme } from '../../../utils/themes'
 
-const useStyles = makeStyles(({ palette, spacing }) => {
+const useStyles = makeStyles<CCFTheme>(({ palette, spacing }) => {
   return {
     title: {
       color: palette.primary.contrastText,