+++ packages/client/src/pages/RecommendationsPage/RecommendationsFilterBar/utils/options/CloudProviderChooser.ts
@@ -2,10 +2,14 @@
  * © 2021 Thoughtworks, Inc.
  */
 
-import { DropdownFilterOptions, DropdownOption, FilterOptions } from 'Types'
-import { DropdownSelections } from 'common/FilterBar/utils/FiltersUtil'
-import { CLOUD_PROVIDER_OPTIONS } from 'common/FilterBar/utils/DropdownConstants'
-import { OptionChooser } from 'common/FilterBar/utils/OptionChooser'
+import {
+  DropdownFilterOptions,
+  DropdownOption,
+  FilterOptions,
+} from '../../../../../Types'
+import { DropdownSelections } from '../../../../../common/FilterBar/utils/FiltersUtil'
+import { CLOUD_PROVIDER_OPTIONS } from '../../../../../common/FilterBar/utils/DropdownConstants'
+import { OptionChooser } from '../../../../../common/FilterBar/utils/OptionChooser'
 
 export class CloudProviderChooser extends OptionChooser {
   constructor(