+++ packages/client/src/pages/RecommendationsPage/RecommendationsPage.test.tsx
@@ -3,28 +3,45 @@
  */
 
 import React from 'react'
-import { fireEvent, render, screen } from '@testing-library/react'
+import { render } from '@testing-library/react'
 import {
   EstimationResult,
   RecommendationResult,
 } from '@cloud-carbon-footprint/common'
 import RecommendationsPage from './RecommendationsPage'
-import { generateEstimations, mockRecommendationData } from 'utils/data'
-import { useRemoteRecommendationsService, useRemoteService } from 'utils/hooks'
-import { ServiceResult } from 'Types'
+import { generateEstimations, mockRecommendationData } from '../../utils/data'
+import {
+  useRemoteFootprintService,
+  useRemoteRecommendationsService,
+} from '../../utils/hooks'
+import { ServiceResult } from '../../Types'
 import moment from 'moment'
-import { act } from 'react-dom/test-utils'
 
-jest.mock('utils/hooks/RecommendationsServiceHook')
-jest.mock('utils/hooks/RemoteServiceHook')
+jest.mock('../../utils/hooks/RecommendationsServiceHook')
+jest.mock('../../utils/hooks/FootprintServiceHook')
+jest.mock('../../ConfigLoader', () => ({
+  __esModule: true,
+  default: () => ({
+    CURRENT_PROVIDERS: [
+      { key: 'aws', name: 'AWS' },
+      { key: 'gcp', name: 'GCP' },
+    ],
+    PREVIOUS_YEAR_OF_USAGE: true,
+    DATE_RANGE: {
+      VALUE: '7',
+      TYPE: 'days',
+    },
+    BASE_URL: '/api',
+  }),
+}))
 
 const mockedUseRecommendationsService =
   useRemoteRecommendationsService as jest.MockedFunction<
     typeof useRemoteRecommendationsService
   >
 
-const mockUseRemoteService = useRemoteService as jest.MockedFunction<
-  typeof useRemoteService
+const mockUseRemoteService = useRemoteFootprintService as jest.MockedFunction<
+  typeof useRemoteFootprintService
 >
 
 // Set Test Date for Moment
@@ -37,6 +54,7 @@ describe('Recommendations Page', () => {
       {
         loading: false,
         data: mockRecommendationData,
+        error: null,
       }
     mockedUseRecommendationsService.mockReturnValue(
       mockRecommendationsReturnValue,
@@ -46,6 +64,7 @@ describe('Recommendations Page', () => {
     const mockReturnValue: ServiceResult<EstimationResult> = {
       loading: false,
       data: data,
+      error: null,
     }
     mockUseRemoteService.mockReturnValue(mockReturnValue)
   })
@@ -68,22 +87,16 @@ describe('Recommendations Page', () => {
   })
 
   it('should pass in to remote service hook today and one month prior', () => {
-    render(<RecommendationsPage />)
+    const onApiError = jest.fn()
+    render(<RecommendationsPage onApiError={onApiError} />)
 
-    const parameters = mockUseRemoteService.mock.calls[0]
+    const parameters = mockUseRemoteService.mock.calls[0][0]
 
-    expect(parameters.length).toEqual(4)
-
-    const initial = parameters[0]
-    const startDate = parameters[1]
-    const endDate = parameters[2]
-    const ignoreCache = parameters[3]
-
-    expect(initial).toEqual([])
-    expect(startDate.month()).toEqual(endDate.month() - 1)
-
-    expect(endDate.isSame(moment.utc(), 'day')).toBeTruthy()
-    expect(ignoreCache).toBeTruthy()
+    expect(parameters.startDate.month()).toEqual(parameters.endDate.month() - 1)
+    expect(parameters.endDate.isSame(moment.utc(), 'day')).toBeTruthy()
+    expect(parameters.ignoreCache).toBeTruthy()
+    expect(parameters.onApiError).toEqual(onApiError)
+    expect(parameters.baseUrl).toEqual('/api')
   })
 
   it('should show loading icon if data has not been returned', () => {
@@ -103,79 +116,6 @@ describe('Recommendations Page', () => {
     const { getByRole } = render(<RecommendationsPage />)
 
     expect(getByRole('progressbar')).toBeInTheDocument()
-  })
-
-  it('displays a selected recommendation in a side panel when its row is clicked', () => {
-    const { getByText, queryByTestId } = render(<RecommendationsPage />)
-
-    expect(queryByTestId('sideBarTitle')).toBeFalsy()
-
-    fireEvent.click(screen.getByText('test-a'))
-    const recommendationDetail = getByText(
-      mockRecommendationData[0].recommendationDetail,
-    )
-
-    expect(queryByTestId('sideBarTitle')).toBeInTheDocument()
-    expect(recommendationDetail).toBeInTheDocument()
-  })
-
-  it('hides and re-displays the side panel when a recommendation row is clicked multiple times', () => {
-    const { getByText, queryByTestId } = render(<RecommendationsPage />)
-
-    fireEvent.click(screen.getByText('test-a'))
-    fireEvent.click(screen.getByText('test-a', { selector: 'div' }))
-    // selector added because now there are multiple elements with the text 'test-a' on the screen
-
-    expect(queryByTestId('sideBarTitle')).toBeFalsy()
-
-    fireEvent.click(screen.getByText('test-a'))
-    const recommendationDetail = getByText(
-      mockRecommendationData[0].recommendationDetail,
-    )
-
-    expect(queryByTestId('sideBarTitle')).toBeInTheDocument()
-    expect(recommendationDetail).toBeInTheDocument()
-  })
-
-  it('displays a new recommendation in the side panel when its recommendation row is clicked', () => {
-    const { getByText, queryByText, queryByTestId } = render(
-      <RecommendationsPage />,
-    )
-
-    fireEvent.click(screen.getByText('test-a'))
-    fireEvent.click(screen.getByText('test-b'))
-    const firstRecommendationDetail = queryByText(
-      mockRecommendationData[0].recommendationDetail,
-    )
-    const secondRecommendationDetail = getByText(
-      mockRecommendationData[1].recommendationDetail,
-    )
-
-    expect(queryByTestId('sideBarTitle')).toBeInTheDocument()
-    expect(firstRecommendationDetail).toBeFalsy()
-    expect(secondRecommendationDetail).toBeInTheDocument()
-  })
-
-  it('re-displays the side panel when the close button and then the same recommendation row is clicked', () => {
-    const { getByText, getByTestId, queryByTestId } = render(
-      <RecommendationsPage />,
-    )
-
-    fireEvent.click(screen.getByText('test-a'))
-
-    const closeIcon = getByTestId('closeIcon')
-
-    act(() => {
-      fireEvent.click(closeIcon)
-    })
-
-    fireEvent.click(screen.getByText('test-a'))
-    const recommendationDetail = getByText(
-      mockRecommendationData[0].recommendationDetail,
-    )
-
-    expect(queryByTestId('sideBarTitle')).toBeInTheDocument()
-    expect(recommendationDetail).toBeInTheDocument()
   })
 
   it('should render a filter bar on the page', () => {